refactor: dedup SEOFooter — fold ContentSEOFooter back into the shared one
ContentSEOFooter existed only because SEOFooter didn't accept a locale.
Promote locale to a prop (defaults 'en'), apply it via the same
localizeHref helper, drop the duplicate. Marketing layout now uses the
unified Footer with showSiteDirectory=true (the default) instead of
disabling it and rendering ContentSEOFooter alongside.

No visual change — both footers rendered the same 4 sections; only the
locale-aware-href behavior carries over.
