+++ src/components/LandingPage/Footer.tsx
@@ -7,7 +7,7 @@ import handPeace from '@/assets/illustrations/hand-peace.svg'
 import handMiddleFinger from '@/assets/illustrations/hand-middle-finger.svg'
 import { SEOFooter } from './SEOFooter'
 
-const Footer = ({ showSiteDirectory = true }: { showSiteDirectory?: boolean }) => {
+const Footer = ({ showSiteDirectory = true, locale = 'en' }: { showSiteDirectory?: boolean; locale?: string }) => {
     return (
         <>
             <footer className="relative flex h-52 items-end justify-between bg-black px-8 pb-10 md:items-center md:px-20 md:pb-0">
@@ -107,7 +107,7 @@ const Footer = ({ showSiteDirectory = true }: { showSiteDirectory?: boolean }) =
                     <Image src={handWaving.src} alt="Hand waving" width={25} height={25} />
                 </section>
             </footer>
-            {showSiteDirectory && <SEOFooter />}
+            {showSiteDirectory && <SEOFooter locale={locale} />}
         </>
     )
 }
