+++ src/components/LandingPage/SEOFooter.tsx
@@ -12,6 +12,20 @@ interface ManifestEntry {
     external?: boolean
 }
 
+function localizeHref(href: string, locale: string): string {
+    if (href.startsWith('/en/')) return `/${locale}/${href.slice(4)}`
+    return href
+}
+
+function FooterSection({ title, children }: { title: string; children: React.ReactNode }) {
+    return (
+        <div>
+            <h3 className="mb-3 text-xs font-bold text-white">{title}</h3>
+            <ul className="space-y-1">{children}</ul>
+        </div>
+    )
+}
+
 function FooterLink({ href, external, children }: { href: string; external?: boolean; children: React.ReactNode }) {
     if (external) {
         return (
@@ -36,7 +50,7 @@ function FooterLink({ href, external, children }: { href: string; external?: boo
     )
 }
 
-export function SEOFooter() {
+export function SEOFooter({ locale = 'en' }: { locale?: string } = {}) {
     const sendTo = (manifest.sendMoney?.to ?? []) as ManifestEntry[]
     const sendFrom = (manifest.sendMoney?.from ?? []) as ManifestEntry[]
     const compare = (manifest.compare ?? []) as ManifestEntry[]
@@ -44,64 +58,54 @@ export function SEOFooter() {
     const resources = (manifest.resources ?? []) as ManifestEntry[]
     const hasSendMoney = sendTo.length > 0 || sendFrom.length > 0
 
+    const link = (entry: ManifestEntry) => (entry.external ? entry.href : localizeHref(entry.href, locale))
+
     return (
         <nav aria-label="Site directory" className="bg-black px-8 py-8 md:px-20">
             <div className="flex flex-wrap justify-between gap-y-8">
                 {hasSendMoney && (
-                    <div>
-                        <h3 className="mb-3 text-xs font-bold text-white">Send Money</h3>
-                        <ul className="space-y-1">
-                            {sendTo.map((entry) => (
-                                <FooterLink key={`to-${entry.slug}`} href={entry.href}>
-                                    Send to {entry.name}
-                                </FooterLink>
-                            ))}
-                            {sendFrom.map((entry) => (
-                                <FooterLink key={`from-${entry.slug}`} href={entry.href}>
-                                    Send from {entry.name}
-                                </FooterLink>
-                            ))}
-                        </ul>
-                    </div>
+                    <FooterSection title="Send Money">
+                        {sendTo.map((entry) => (
+                            <FooterLink key={`to-${entry.slug}`} href={link(entry)}>
+                                Send to {entry.name}
+                            </FooterLink>
+                        ))}
+                        {sendFrom.map((entry) => (
+                            <FooterLink key={`from-${entry.slug}`} href={link(entry)}>
+                                Send from {entry.name}
+                            </FooterLink>
+                        ))}
+                    </FooterSection>
                 )}
 
                 {compare.length > 0 && (
-                    <div>
-                        <h3 className="mb-3 text-xs font-bold text-white">Compare</h3>
-                        <ul className="space-y-1">
-                            {compare.map((entry) => (
-                                <FooterLink key={entry.slug} href={entry.href}>
-                                    Peanut vs {entry.name}
-                                </FooterLink>
-                            ))}
-                        </ul>
-                    </div>
+                    <FooterSection title="Compare">
+                        {compare.map((entry) => (
+                            <FooterLink key={entry.slug} href={link(entry)}>
+                                Peanut vs {entry.name}
+                            </FooterLink>
+                        ))}
+                    </FooterSection>
                 )}
 
                 {articles.length > 0 && (
-                    <div>
-                        <h3 className="mb-3 text-xs font-bold text-white">Learn More</h3>
-                        <ul className="space-y-1">
-                            {articles.map((entry) => (
-                                <FooterLink key={entry.slug} href={entry.href}>
-                                    {entry.name}
-                                </FooterLink>
-                            ))}
-                        </ul>
-                    </div>
+                    <FooterSection title="Learn More">
+                        {articles.map((entry) => (
+                            <FooterLink key={entry.slug} href={link(entry)}>
+                                {entry.name}
+                            </FooterLink>
+                        ))}
+                    </FooterSection>
                 )}
 
                 {resources.length > 0 && (
-                    <div>
-                        <h3 className="mb-3 text-xs font-bold text-white">Resources</h3>
-                        <ul className="space-y-1">
-                            {resources.map((entry) => (
-                                <FooterLink key={entry.slug} href={entry.href} external={entry.external}>
-                                    {entry.name}
-                                </FooterLink>
-                            ))}
-                        </ul>
-                    </div>
+                    <FooterSection title="Resources">
+                        {resources.map((entry) => (
+                            <FooterLink key={entry.slug} href={link(entry)} external={entry.external}>
+                                {entry.name}
+                            </FooterLink>
+                        ))}
+                    </FooterSection>
                 )}
             </div>
         </nav>
